+++ src/lib.rs
@@ -39,4 +39,4 @@ pub use error::{Error, ErrorKind, Result, ResultExt};
 pub use identifiable::Identifiable;
 pub use sized::MtProtoSized;
 pub use ser::{Serializer, to_bytes, to_writer};
-pub use de::{Deserializer, from_bytes, from_reader};
+pub use de::{Deserializer, from_bytes, from_bytes_reuse, from_reader, from_reader_reuse};